+++ Adjust/ADJAdjust2dx.mm
@@ -32,8 +32,8 @@ void ADJAdjust2dx::setEnabled(bool isEnabled) {
     [Adjust setEnabled:isEnabled];
 }
 
-void ADJAdjust2dx::appWillOpenUrl(void* url) {
-    NSURL *pUrl = (NSURL *)url;
+void ADJAdjust2dx::appWillOpenUrl(std::string url) {
+    NSURL *pUrl = [NSURL URLWithString:[NSString stringWithUTF8String:url.c_str()]];
     [Adjust appWillOpenUrl:pUrl];
 }
 
@@ -46,6 +46,35 @@ void ADJAdjust2dx::setDeviceToken(std::string deviceToken) {
 void ADJAdjust2dx::setOfflineMode(bool isOffline) {
     [Adjust setOfflineMode:isOffline];
 }
+
+void ADJAdjust2dx::sendFirstPackages() {
+    [Adjust sendFirstPackages];
+}
+
+void ADJAdjust2dx::addSessionCallbackParameter(std::string key, std::string value) {
+    [Adjust addSessionCallbackParameter:[NSString stringWithUTF8String:key.c_str()] value:[NSString stringWithUTF8String:value.c_str()]];
+}
+
+void ADJAdjust2dx::addSessionPartnerParameter(std::string key, std::string value) {
+    [Adjust addSessionPartnerParameter:[NSString stringWithUTF8String:key.c_str()] value:[NSString stringWithUTF8String:value.c_str()]];
+}
+
+void ADJAdjust2dx::removeSessionCallbackParameter(std::string key) {
+    [Adjust removeSessionCallbackParameter:[NSString stringWithUTF8String:key.c_str()]];
+}
+
+void ADJAdjust2dx::removeSessionPartnerParameter(std::string key) {
+    [Adjust removeSessionPartnerParameter:[NSString stringWithUTF8String:key.c_str()]];
+}
+
+void ADJAdjust2dx::resetSessionCallbackParameters() {
+    [Adjust resetSessionCallbackParameters];
+}
+
+void ADJAdjust2dx::resetSessionPartnerParameters() {
+    [Adjust resetSessionPartnerParameters];
+}
+
 
 bool ADJAdjust2dx::isEnabled() {
     return [Adjust isEnabled];